move where sources are located to an unified place
