+++ martin/src/pg/query_tables.rs
@@ -14,7 +14,7 @@ use crate::config::file::pg::{PgInfo, TableInfo};
 use crate::pg::PgError::PostgresError;
 use crate::pg::PgResult;
 use crate::pg::builder::SqlTableInfoMapMapMap;
-use crate::pg::pg_source::PgSqlInfo;
+use crate::pg::source::PgSqlInfo;
 use crate::pg::pool::PgPool;
 use crate::pg::utils::{json_to_hashmap, polygon_to_bbox};
